requirements.txt
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==2.2.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-dotenv==1.0.0
python-telegram-bot==20.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.1.3
